Move splitVcsPathFromUrl() from Utils to VersionControlSystem
This way we can return an VcsInfo object which was not possible before
the move as the "util" module must not depend on the "model" module.

Returning a VcsInfo has the advantage of also being able to return more
VCS information contained in the URL, like the revision. Also, we can
leverage the VersionControlSystem function to detect the provider for
arbitrary URLs.
